requirements.txt
asgiref==3.4.1
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-summernote==0.8.20.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
facebook-sdk==3.1.0
future==0.18.2
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
oauthlib==3.1.1
packaging==21.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==2.4.7
python-dotenv==0.19.2
python-http-client==3.3.3
python-twitter==3.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sendgrid==6.9.0
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
